requirements.txt
playwright==1.44.0
python-telegram-bot==21.3
python-dotenv==1.0.1
ollama==0.2.1
httpx==0.27.0
aiosqlite==0.20.0
